Fix some CloudFront Distribution defaults
Some of the CloudFront Distribution resource schema defaults are incorrect,
per https://github.com/terraform-providers/terraform-provider-aws/issues/1659.
This was a minor annoyance for the TTL fields (whose defaults I've taken from
CloudFormation); however, in the case of restrictions, we actually can't specify
what we need to specify.  The logic in the provider already works conditionally
based on its presence, so all we need to do is tweak Optional vs. Required.