Add retries for resource provisioning
In pulumi/pulumi-terraform#56, we are seeing resources get created
but their describe calls turn up empty handed.  Although we haven't
validated that the calls would eventually, in fact, find the resource,
it makes sense that immediately calling describe may suffer from races
on the AWS backend where during a minimal window of time the newly
provisioned resource wouldn't exist.  Many resource providers have
retry loops for this very reason.  But not all of them ...

We've been bitten by ECR repositories and ELBv2 listeners
disproportionately recently.  I'm sure this is because we just happen
to be hammering these harder these days thanks to all of the container
test cases.  So, this fix adds retry loops to just those two resource
types, but before calling it quits we'll need to do a more exhaustive
audit of the other resource types to fix up more of them.  Essentially
any resource that just returns the result of xxRead as a tail call
from its xxCreate method is likely suspect.