aws_ecs_service: Honour wait_for_steady_state on create
Previously, wait_for_steady_state was only honoured for updates. Note that we
set the ID before waiting for the service to come to a steady state such that we
don't lose the ID in the event of failure.

Additionally, the ECS `DescribeServices` API doesn't return
`ServiceNotFoundException` when provided with a nonexistent service ID. Instead,
it returns a failure for the corresponding ARN with the reason `MISSING`. We
retry a number of times since it is more likely we hit this than that the
service was deleted underneath us.