Increase timeout for ECS Service wait_for_steady_state
The timeout wwas previously set to 10m to match https://docs.aws.amazon.com/sdk-for-go/api/service/ecs/#ECS.WaitUntilServicesStable.  However, it is very common for ECS services to take just over 10 minutes to get to a steady state after an update, particularly if they are attached to a load balancer - which defaults to just about 10 minutes of waiting for healthy and waiting to drain old tasks during an update.