feat: Allow removal of EKS cluster addons
​
This commit adds a new parameter `default_addons_to_remove` to the
EKS cluster API, which takes a list of addons to be removed. If set the
addon will first be installed, which will adopt any unmanaged, preinstalled
addon and bring it under management of Terraform. Afterwards it will be
deleted, which will fully remove the addon from EKS.
​
This will allow any addons to that come preinstalled with an EKS cluster
to be removed upon creation and will allow them to be fully managed via
e.g. Pulumi Kubernetes resources.