Revert framework conversions
This commit reverts resources and data sources that were ported to the
plugin framework, back to their previous implementation. The upstream
implementations are kept in their files, but with an `ignores` build tag
so they won't be compiled. This means changes upstream are much less
likely to cause conflicts. The reverted implementations are in the files
*_reverted.go.