Reverts patches to S3BucketLegacy and GameLift
Previously, we were pulling along patches which removed a few simplifications
to waiters in AWS GameLift, and a newer patch which plumbed through context.Context
in various places for CRUD operations in S3 Buckets (which are now ported to Legacy)
and in GameLift - specifically matchmaking configurations and rulesets.

This PR reverts our patches which will prevent future merge conflicts, and updates
call sites to plumb through the proper Context objects.