Marks SSE configuration as Computed for S3BucketLegacy
In January 2023, AWS turned SSE on by default for all new S3 buckets.
This change was rolled into the TF provider for the new version of S3
bucket, but since we maintain a fork which contains the legacy bucket as well,
we need to incorporate that change in both places.