Marks SSE Configuration as Computed for S3BucketLegacy
Previously, we assumed that AWS would not make changes to the legacy
s3 bucket resource, but in January SSE was enabled by default for all
new s3 buckets.  This change was backported into the terraform provider
but as we maintain a fork, which includes a resource not maintained by
AWS, we did not propagate that change here.