Marks SSE Configuration as Computed for Legacy S3 Bucket
In January, AWS enabled SSE by default for all new S3 buckets.
Since we maintain a forked version of the legacy S3 bucket resource,
this change was not properly propagated to the legacy S3 bucket, which
appears in pulumi as `aws.s3.Bucket` and is likely the default version of
the Bucket resource that users will use.